wip/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-jose
passlib
